+++ ptr_table.h
@@ -31,6 +31,10 @@ struct _ptr_record {
 	void* address;  /* value1 */
 	PtrType type; /* value2 */
 	GCReq gc; /* value3: whether garbage collected. 0: GC_NO ,  1: GC_YES */
+	/* Mainly used for numbers which requires memory for double and int. */
+	void* ex_addr; 
+	PtrType ex_type;
+	GCReq ex_gc; 
 	UT_hash_handle hh; /* This macro makes this structure hashable */
 };
 typedef struct _ptr_record ptr_record ;
@@ -41,12 +45,15 @@ ptr_table*  ptr_table_init();
 ptr_record* ptr_table_add(ptr_table** table, char* key, void** address, PtrType type, GCReq gc);
 
 ptr_record* ptr_table_create_int(ptr_table** table, char* key, int ival);
-ptr_record* ptr_table_create_int_from_ptr(ptr_table** table, char* key, int** iptr);
+ptr_record* ptr_table_create_int_from_ptr(ptr_table** table, char* key, int** iptr, double** dptr);
 int ptr_table_update_int(ptr_table** table, char* key, int ival);
 
 ptr_record* ptr_table_create_double(ptr_table** table, char* key, double dval);
-ptr_record* ptr_table_create_double_from_ptr(ptr_table** table, char* key, double** dptr);
+ptr_record* ptr_table_create_double_from_ptr(ptr_table** table, char* key, double** dptr, int** iptr);
 int ptr_table_update_double(ptr_table** table, char* key, double dval);
+
+int ptr_record_update_extra_address(ptr_record* pr, void** ptr_ex_addr, PtrType ex_type, GCReq ex_gc );
+int ptr_record_swap_addresses(ptr_record* pr);
 
 /*
 ptr_record* ptr_table_create_anonym_struct_string(ptr_table** table, struct_string** strptr);
@@ -56,6 +63,7 @@ ptr_record* ptr_table_create_anonym_string(ptr_table** table, string_object** st
 ptr_record* ptr_table_create_string(ptr_table** table, char* key, string_object** strptr);
 ptr_record* ptr_table_create_string_from_ptr(ptr_table** table, char* key, string_object** strptr);
 int ptr_table_update_stirng(ptr_table**, char* key, string_object** strptr);
+
 
 string_object* ptr_table_get_pp_string(ptr_table** table, char* key);
 const char* ptr_table_get_char(ptr_table** table, char* key);